Achieved successful compillation. To do tests.

+++ src/main/scala/com/bwsw/tstreams/agents/group/CheckpointInfo.scala
@@ -34,7 +34,7 @@ case class ProducerCheckpointInfo(transactionRef: ProducerTransaction,
                                   partition: Int,
                                   transaction: Long,
                                   totalCnt: Int,
-                                  ttl: Int) extends CheckpointInfo
+                                  ttl: Long) extends CheckpointInfo
 
 /**
   * BasicConsumer commit information
